add docs for helper functions
