+++ src/bundle/processor/index.ts
@@ -1,0 +1,1 @@
+export * from './bundle-processor.js'
